+++ website/scripts/changelog.md
@@ -1,2 +1,15 @@
+# changes
+
 - `/intro/getting-started/install` redirects to `/intro/getting-started`
 - `/guides/hcl` folders with only index changed to named individual files
+- `/docs/install/index.html` to `docs/install.html`
+- `/docs/other` dumped to root
+- `/docs/basics/terminology` -> `/docs/terminology`
+- `/docs/configuration/from-1.5` -> `/docs/from-1.5`
+- `/docs/from-1.5/functions.html` -> `/docs/from-1.5/functions/index.html`
+- [BREAKING] `/docs/from-1.5/functions/collection/index.html` -> `/docs/from-1.5/functions/collection/index-fn.html`
+- `/docs/from-1.5/functions/*/overview.html` -> `/docs/from-1.5/functions/*/index.html`
+
+# problems:
+
+- empty index files on all `from-1.5/functions/*`